+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.6 - 2024-07-12
+
+**Added:** Backup date meta field
+**Updated:** Cron job for auth token status
+
 ## v1.0.5 - 2024-07-12
 
 **Updated:** Fixes in cron job `wpc2_gdoc_register_run_gdrive_backup_job` about post status and pagination